+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/activity/UserDetailActivity.kt
@@ -13,11 +13,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.TaskStackBuilder
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.*
-import androidx.recyclerview.widget.DiffUtil
-import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
 import dagger.hilt.android.AndroidEntryPoint
@@ -38,7 +34,6 @@ import net.pantasystem.milktea.common_android_ui.UserPinnedNotesFragmentFactory
 import net.pantasystem.milktea.common_android_ui.report.ReportDialog
 import net.pantasystem.milktea.common_navigation.*
 import net.pantasystem.milktea.common_viewmodel.confirm.ConfirmViewModel
-import net.pantasystem.milktea.model.account.page.Pageable
 import net.pantasystem.milktea.model.setting.Config
 import net.pantasystem.milktea.model.setting.LocalConfigRepository
 import net.pantasystem.milktea.model.setting.Theme
@@ -52,10 +47,9 @@ import net.pantasystem.milktea.user.databinding.ActivityUserDetailBinding
 import net.pantasystem.milktea.user.nickname.EditNicknameDialog
 import net.pantasystem.milktea.user.profile.ConfirmUserBlockDialog
 import net.pantasystem.milktea.user.profile.ProfileAccountSwitchDialog
+import net.pantasystem.milktea.user.profile.ProfileTabPagerAdapter
 import net.pantasystem.milktea.user.profile.UserProfileFieldListAdapter
 import net.pantasystem.milktea.user.profile.mute.SpecifyMuteExpiredAtDialog
-import net.pantasystem.milktea.user.reaction.UserReactionsFragment
-import net.pantasystem.milktea.user.viewmodel.UserDetailTabType
 import net.pantasystem.milktea.user.viewmodel.UserDetailViewModel
 import javax.inject.Inject
 
@@ -209,7 +203,7 @@ class UserDetailActivity : AppCompatActivity() {
         )
             .initViewModelListener()
 
-        val adapter = UserTimelinePagerAdapterV2(
+        val adapter = ProfileTabPagerAdapter(
             pageableFragmentFactory,
             userPinnedNotesFragmentFactory,
             this
@@ -472,111 +466,3 @@ class UserDetailActivity : AppCompatActivity() {
     }
 }
 
-class UserTimelinePagerAdapterV2(
-    val pageableFragmentFactory: PageableFragmentFactory,
-    val userPinnedNotesFragmentFactory: UserPinnedNotesFragmentFactory,
-    activity: FragmentActivity,
-) : FragmentStateAdapter(activity) {
-
-    var tabs: List<UserDetailTabType> = emptyList()
-        private set
-
-    override fun createFragment(position: Int): Fragment {
-        return when (val tab = tabs[position]) {
-            is UserDetailTabType.Gallery -> pageableFragmentFactory.create(
-                tab.accountId,
-                Pageable.Gallery.User(tab.userId.id),
-            )
-            is UserDetailTabType.Media -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.UserTimeline(
-                    tab.userId.id,
-                    withFiles = true
-                )
-            )
-            is UserDetailTabType.PinNote -> userPinnedNotesFragmentFactory.create(tab.userId)
-            is UserDetailTabType.Reactions -> UserReactionsFragment.newInstance(tab.userId)
-            is UserDetailTabType.UserTimeline -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.UserTimeline(
-                    tab.userId.id,
-                    includeReplies = false
-                )
-            )
-            is UserDetailTabType.UserTimelineWithReplies -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.UserTimeline(
-                    tab.userId.id,
-                    includeReplies = true
-                )
-            )
-            is UserDetailTabType.MastodonMedia -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.Mastodon.UserTimeline(
-                    tab.userId.id,
-                    isOnlyMedia = true,
-                )
-            )
-            is UserDetailTabType.MastodonUserTimeline -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.Mastodon.UserTimeline(
-                    tab.userId.id,
-                    excludeReplies = true,
-                )
-            )
-            is UserDetailTabType.MastodonUserTimelineWithReplies -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.Mastodon.UserTimeline(
-                    tab.userId.id,
-                    excludeReplies = false,
-                )
-            )
-            is UserDetailTabType.MastodonUserTimelineOnlyPosts -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.Mastodon.UserTimeline(
-                    tab.userId.id,
-                    excludeReblogs = true,
-                )
-            )
-            is UserDetailTabType.UserTimelineOnlyPosts -> pageableFragmentFactory.create(
-                tab.userId.accountId,
-                Pageable.UserTimeline(
-                    tab.userId.id,
-                    includeMyRenotes = false,
-                ),
-            )
-        }
-
-    }
-
-    override fun getItemCount(): Int {
-        return tabs.size
-    }
-
-    fun submitList(list: List<UserDetailTabType>) {
-
-        val old = tabs
-        tabs = list
-        val callback = object : DiffUtil.Callback() {
-            override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-                return old[oldItemPosition] == list[newItemPosition]
-            }
-
-            override fun getNewListSize(): Int {
-                return list.size
-            }
-
-            override fun getOldListSize(): Int {
-                return old.size
-            }
-
-            override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-                return old[oldItemPosition] == list[newItemPosition]
-            }
-        }
-        val result = DiffUtil.calculateDiff(callback)
-        result.dispatchUpdatesTo(this)
-    }
-
-
-}
